Added the String helpers chars() and bytes().
The String helper `chars` (aliased as `each_char`) passes each character of the **string** to the given **iterator**, or returns an iterable if no iterator is given.

The String helper `bytes` (aliased as `each_byte`) passes each byte in **string** to the given **iterator**, or returns an iterable if no iterator is given.