Added the String helper count().
The String helper count counts the number of occurrences defined by a set of characters given in extra parameters. The intersection of these sets defines the characters to count in the given string.
Any other parameter that starts with a caret (`^`) is negated.
The sequence `c1–c2` means all characters between `c1` and `c2`.

A port of the native Ruby `String#count` function. See the Ruby documentation :
http://ruby-doc.org/core-1.9.3/String.html#method-i-count